3.3_permissions/api_with_restrictions/requirements-dev.txt
-r requirements.txt
pytest-django
model_bakery
